+++ src/index.ts
@@ -386,7 +386,6 @@ export type {
 } from './runtime/IAtcRun';
 export type {
   ICrossTrace,
-  ICrossTraceDocuments,
   ICrossTraceResults,
   IListCrossTracesOptions,
 } from './runtime/ICrossTrace';
@@ -438,7 +437,6 @@ export type {
   IGenerateServiceBindingParams,
   IGetServiceBindingODataParams,
   IPublishODataV2Params,
-  IServiceBindingDocuments,
   IServiceBindingResults,
   ITransportCheckServiceBindingParams,
   IUnpublishODataV2Params,